Add some xml doc comments to new classes. Move your new classes under the WebSocketSharp.Net namespace. And move ServerSslAuthConfiguration.cs to Net folder. Stay ToolsVersion="3.5" in websocket-sharp.csproj
